Refactor acceptance tests, combine kafka user and topic management together.
