+++ acc_test/data/kafka_topic_topic_list.tf
@@ -1,4 +1,4 @@
-// This is part of testing "kafka topic" suite, 3 of 5
+// This is part of testing "kafka topic" suite, 2 of 4
 provider "instaclustr" {
   username = "%s"
   api_key = "%s"
